backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.2
stripe==7.6.0
python-dotenv==1.0.0
asyncpg==0.29.0
starlette==0.27.0
websockets==12.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
email-validator==2.1.0
loguru==0.7.2
